Bug fixed register screen profile picture encoding, Added delete account and reset password (To Do: Edit Profile, many UI improvements)

+++ core/domain/src/main/java/com/example/domain/user/User.kt
@@ -1,12 +1,14 @@
 package com.example.domain.user
 
+import java.math.BigInteger
+
 data class User(
     val user_id: Int,
     val first_name: String,
     val last_name: String,
     val username: String,
     val gender: String,
-    val mobile_number: Int,
+    val mobile_number: Long,
     val email: String,
     val address: String,
     val profile_pic: String,
